enable_live_data.py
#!/usr/bin/env python3
"""
Enable Live Data Sources for Australian Company Pipeline
This script shows how to connect to real Common Crawl and ABR data sources.
"""

import asyncio
import aiohttp
import xml.etree.ElementTree as ET
from typing import List, Dict, Any
import logging

logger = logging.getLogger(__name__)

class LiveCommonCrawlExtractor:
    """
    Real Common Crawl extractor for Australian company websites.
    Connects to commoncrawl.org APIs and processes website data.
    """
    
    def __init__(self):
        self.base_url = "https://commoncrawl.org"
        self.index_url = "https://index.commoncrawl.org"
        
    async def get_latest_crawl_index(self) -> str:
        """Get the latest Common Crawl index identifier."""
        async with aiohttp.ClientSession() as session:
            async with session.get(f"{self.base_url}/crawl-data/") as response:
                # Parse HTML to find latest crawl (e.g., CC-MAIN-2025-03)
                html = await response.text()
                # Implementation would parse for latest index
                return "CC-MAIN-2025-03"  # March 2025 example
    
    async def query_australian_domains(self, index_name: str) -> List[str]:
        """Query Common Crawl index for Australian domains."""
        australian_tlds = ['.com.au', '.net.au', '.org.au', '.edu.au', '.gov.au', '.asn.au']
        
        urls = []
        async with aiohttp.ClientSession() as session:
            for tld in australian_tlds:
                query_url = f"{self.index_url}/CC-MAIN-{index_name}-index"
                params = {
                    'url': f'*{tld}/*',
                    'output': 'json',
                    'limit': 10000  # Adjust based on needs
                }
                
                try:
                    async with session.get(query_url, params=params) as response:
                        if response.status == 200:
                            data = await response.text()
                            # Parse JSONL response
                            for line in data.strip().split('\n'):
                                if line:
                                    record = eval(line)  # In production, use json.loads
                                    urls.append(record.get('url'))
                        
                except Exception as e:
                    logger.error(f"Error querying {tld}: {e}")
                    
        return urls[:200000]  # Limit to target 200k websites
    
    async def extract_company_data(self, url: str) -> Dict[str, Any]:
        """Extract company information from website content."""
        async with aiohttp.ClientSession() as session:
            try:
                async with session.get(url, timeout=10) as response:
                    if response.status == 200:
                        html = await response.text()
                        
                        # Use LLM to extract company information
                        # (This would integrate with your existing LLM client)
                        return {
                            'website_url': url,
                            'company_name': 'Extracted via LLM',
                            'industry': 'Classified via LLM',
                            'content_snippet': html[:1000]
                        }
                        
            except Exception as e:
                logger.error(f"Error extracting from {url}: {e}")
                
        return None

class LiveABRExtractor:
    """
    Real ABR extractor that processes bulk XML files from data.gov.au
    """
    
    def __init__(self):
        self.abr_base_url = "https://data.gov.au"
        self.bulk_extract_url = "https://data.gov.au/data/dataset/australian-business-register-bulk-extract"
    
    async def download_abr_bulk_file(self) -> str:
        """Download the latest ABR bulk extract XML file."""
        # Implementation would:
        # 1. Check data.gov.au for latest bulk extract
        # 2. Download large ZIP file (several GB)
        # 3. Extract XML files
        # 4. Return path to XML file
        
        print("🔄 Would download ABR bulk extract from data.gov.au")
        print("⚠️  Note: ABR files are very large (10+ GB)")
        return "path/to/abr_bulk_extract.xml"
    
    async def process_abr_xml_stream(self, xml_file_path: str) -> List[Dict[str, Any]]:
        """Process ABR XML file using streaming to handle large files."""
        companies = []
        
        # Streaming XML parser for large files
        context = ET.iterparse(xml_file_path, events=('start', 'end'))
        context = iter(context)
        event, root = next(context)
        
        for event, elem in context:
            if event == 'end' and elem.tag == 'ABR_Entity':  # Adjust tag name
                company = self.parse_abr_entity(elem)
                if company:
                    companies.append(company)
                    
                # Clear element to save memory
                elem.clear()
                root.clear()
                
                # Limit for demo
                if len(companies) >= 1000000:
                    break
                    
        return companies
    
    def parse_abr_entity(self, elem) -> Dict[str, Any]:
        """Parse individual ABR entity from XML."""
        return {
            'abn': elem.find('.//ABN').text if elem.find('.//ABN') is not None else None,
            'entity_name': elem.find('.//EntityName').text if elem.find('.//EntityName') is not None else None,
            'entity_type': elem.find('.//EntityType').text if elem.find('.//EntityType') is not None else None,
            'entity_status': elem.find('.//EntityStatus').text if elem.find('.//EntityStatus') is not None else None,
            'address': self.parse_address(elem),
            'start_date': elem.find('.//StartDate').text if elem.find('.//StartDate') is not None else None
        }
    
    def parse_address(self, elem) -> Dict[str, str]:
        """Parse address information from ABR XML."""
        address_elem = elem.find('.//Address')
        if address_elem is not None:
            return {
                'line_1': address_elem.find('.//AddressLine1').text if address_elem.find('.//AddressLine1') is not None else None,
                'suburb': address_elem.find('.//Suburb').text if address_elem.find('.//Suburb') is not None else None,
                'state': address_elem.find('.//State').text if address_elem.find('.//State') is not None else None,
                'postcode': address_elem.find('.//Postcode').text if address_elem.find('.//Postcode') is not None else None
            }
        return {}

async def enable_live_data_extraction():
    """
    Demonstrate how to enable live data extraction from real sources.
    """
    print("🌐 Enabling Live Data Sources for Australian Company Pipeline")
    print("=" * 70)
    
    print("1. Common Crawl Integration:")
    print("   📋 Steps needed:")
    print("   • Connect to commoncrawl.org API")
    print("   • Query March 2025 index for Australian domains")
    print("   • Download and parse ~200,000 websites")
    print("   • Extract company info using GPT-4 Turbo")
    print()
    
    print("2. ABR Integration:")
    print("   📋 Steps needed:")
    print("   • Download bulk XML from data.gov.au")
    print("   • Stream process large files (10+ GB)")
    print("   • Extract ~1M+ business registrations")
    print("   • Validate ABN checksums")
    print()
    
    print("⚠️  Important Considerations:")
    print("   • Common Crawl: Requires significant bandwidth and processing")
    print("   • ABR Data: Very large files, need streaming processing")
    print("   • Rate Limits: Both sources have usage restrictions")
    print("   • Legal: Check terms of service for commercial use")
    print("   • Storage: Need substantial disk space for raw data")
    print()
    
    print("🚀 Current Status:")
    print("   ✅ Pipeline architecture ready")
    print("   ✅ Entity matching system operational")
    print("   ✅ GPT-4 Turbo configured")
    print("   ✅ CSV export system working")
    print("   ⏳ Live data connections not implemented")
    print()
    
    print("💡 Recommendation:")
    print("   Start with smaller test datasets before full 200k/1M+ scale")
    print("   Consider API rate limits and costs")
    print("   Test with sample data first (current 5000 sample works perfectly)")

if __name__ == "__main__":
    asyncio.run(enable_live_data_extraction())